requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
faiss-cpu
pypdf